Fix issues with Nil case in ListDemand
There were two issues stopping the proper detection of strictness on the
`[]` case of a list.

1: The data structure representing demand assumed that the recursive call of a list
was only ever 'Cons' shaped if it was indeed demanded

2: Most functions over list (including `take n`) don't share the `[]` at the
end of a list, instead replacing it with a new one. So even if the demand structure
was correct, you'd never see the instrumented `[]`